+++ RecipesPWA/Common/Components/RecipeForm.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Components.Forms;
-using RecipesPWA.Common.Recipes;
+﻿using RecipesPWA.Common.Recipes;
 
 namespace RecipesPWA.Common.Components;
