+++ RecipesPWA/Shared/FormComponents/IngredientForm.razor
@@ -1,5 +1,5 @@
 ﻿@using RecipesPWA.Common.Recipes
-@inherits GenericForm<Ingredient>
+@inherits RecipesPWA.Common.Components.GenericForm<Ingredient>
 
 <EditForm
           EditContext="@_editContext"
