+++ RecipesPWA/wwwroot/css/app.css
@@ -104,7 +104,6 @@ html, body {
 
 .text-input-area {
     width: 100%;
-    height: 100%;
     overflow: hidden;
     resize: both;
 }
